+++ .env
@@ -1,1 +1,1 @@
-MONGODB_URL = "mongodb+srv://admin:bookstorems@bookstorems.qgl1qca.mongodb.net"
+MONGODB_URL = "mongodb+srv://user1:bookstoreUser1@bookstorems.qgl1qca.mongodb.net"
