+++ package.json
@@ -10,8 +10,10 @@
   },
   "dependencies": {
     "@next/font": "13.1.6",
+    "bootstrap": "^5.3.0-alpha1",
     "eslint": "8.34.0",
     "eslint-config-next": "13.1.6",
+    "mongoose": "^7.0.2",
     "next": "13.1.6",
     "react": "18.2.0",
     "react-dom": "18.2.0"
